requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
mangum==0.17.0
boto3==1.34.34
cursor-ai==0.2.0  # This is a placeholder version
windsurf-mcp==0.1.0  # This is a placeholder version 